requirements.txt
# openai 
# gtts 
# SpeechRecognition 
# # pyaudio 
# streamlit
# python-dotenv


# streamlit openai gtts SpeechRecognition python-dotenv
# pip install 
streamlit 
sounddevice 
numpy 
speechrecognition 
gtts 
python-dotenv 
openai